python_app/requirements.txt
flask==2.3.3
flask-socketio==5.3.2
eventlet==0.33.3
opencv-python-headless==4.8.1.78
mediapipe==0.10.20
mysql-connector-python